added progress bar
