+++ large_test.config
@@ -1,6 +1,6 @@
 server {
     server_name localhost;
-    port 5001;
+    port 8000;
     client_max_body_size 10m;
 
     location / {
@@ -45,6 +45,194 @@ server {
     port 600;
     port 2135;
     port 8080;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 8001;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 5003;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 601;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+server {
+    server_name localhost;
+    port 8005;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 5007;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 606;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+server {
+    server_name localhost;
+    port 80;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 50;
+    client_max_body_size 10m;
+
+    location / {
+        root /var/www/html;
+        index index.html;
+        listing on;
+        allowed_methods GET POST;
+    }
+
+    location /cgi-bin/ {
+        cgi /usr/lib/cgi-bin; #help
+    }
+
+    location /redir/ {
+        redir http://example.com;
+    }
+}
+
+server {
+    server_name localhost;
+    port 21;
+
     client_max_body_size 10m;
 
     location / {
